fix:refactored predictions route
